Feature EPL-1728: Changed the type routes
